Content_Scrapping/requirements.txt
streamlit
pypdf2
langchain
python-dotenv
faiss-cpu
openai
huggingface_hub
#external empadding 
instructorEmbedding
sentence_transeformers